+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Test flag that prevents database interaction and upload to backblaze
+- Add protections that verify certain fields have a value
 
 ### Changed
 - Updated to match most recent Seeking Alpha API changes